implement useSiweAuth hook
